Refactor a couple methods into the beginning of a data.py for classifiers
